Modularize web app script and hydrate home state
