+++ src/app/web/static/translations.js
@@ -1,0 +1,666 @@
+export const translations = {
+  en: {
+    "badge.beta": "Beta",
+    "nav.home": "Home",
+    "nav.profiles": "Profiles",
+    "nav.events": "Events",
+    "nav.rosters": "Rosters",
+    "nav.federations": "Federations",
+    "nav.about": "About",
+    "locale.label": "Choose language",
+    "action.login": "Log in",
+    "action.signup": "Create account",
+    "footer.powered": "Trackeo · Powered by FastAPI · API base:",
+    "footer.location": "Headquartered in Atlanta, GA · Scaling across South America with localized partners.",
+    "hero.eyebrow": "Latin American Athletics Intelligence",
+    "hero.title": "Trackeo connects federations, coaches, and fans across the Americas.",
+    "hero.description":
+      "Explore verified performances, multilingual news, and premium insights that highlight the rise of track & field from São Paulo to Bogotá. Built in Atlanta for the continent.",
+    "hero.explore": "Explore as guest",
+    "hero.subscribe": "View coach plans",
+    "preview.title": "Regional spotlight",
+    "preview.subtitle": "19 federations streaming live splits and heat sheets.",
+    "preview.metric_meets": "Verified meets",
+    "preview.metric_athletes": "Athletes tracked",
+    "preview.metric_rosters": "Roster updates",
+    "search.title": "Search Trackeo's universe",
+    "search.subtitle": "Find athletes, events, rosters, and bilingual news in one place.",
+    "search.placeholder": "Search for an athlete, club, or meet",
+    "search.aria": "Search Trackeo",
+    "search.filters": "Search categories",
+    "search.filter_all": "All",
+    "search.filter_athletes": "Athletes",
+    "search.filter_events": "Events",
+    "search.filter_rosters": "Rosters",
+    "search.filter_news": "News",
+    "search.empty": "Start typing to explore Trackeo's data universe.",
+    "search.no_results": "No matches yet. Try adjusting your filters or spelling.",
+    "home.athletes_title": "Athletes",
+    "home.athletes_subtitle": "Browse the roster or showcase a rising star from your federation.",
+    "home.athletes_seed": "Reload Sample Athletes",
+    "home.athletes_form_title": "Create athlete profile",
+    "home.athletes_submit": "Register athlete",
+    "home.athletes_empty": "No athletes have been registered yet.",
+    "home.athletes_hint": "Use the creation form or load curated sample athletes to populate the roster.",
+    "home.events_title": "Events",
+    "home.events_subtitle": "Track the latest competitions from Atlanta to Buenos Aires.",
+    "home.events_seed": "Reload Sample Events",
+    "home.events_form_title": "Schedule a new meet",
+    "home.events_submit": "Create event",
+    "home.events_empty": "No events have been scheduled yet.",
+    "home.events_hint": "Use the meet form or reload the curated calendar of sample events.",
+    "event_detail.eyebrow": "Live meet operations",
+    "event_detail.refresh": "Refresh",
+    "event_detail.generate_demo": "Generate demo timeline",
+    "event_detail.overview_title": "Event overview",
+    "event_detail.overview_subtitle": "Track status, federation ownership, and last update.",
+    "event_detail.summary_location": "Location",
+    "event_detail.summary_dates": "Dates",
+    "event_detail.summary_federation": "Federation",
+    "event_detail.summary_status": "Live status",
+    "event_detail.schedule_title": "Sessions & timetable",
+    "event_detail.schedule_subtitle": "Monitor blocks of competition and venues.",
+    "event_detail.schedule_empty": "No sessions have been scheduled yet.",
+    "event_detail.schedule_hint": "Create a session above or generate demo data to populate the timeline.",
+    "event_detail.disciplines_title": "Events & scoreboards",
+    "event_detail.disciplines_subtitle": "Heat sheets update instantly for track-side ops.",
+    "event_detail.disciplines_empty": "No disciplines are attached to this meet yet.",
+    "event_detail.disciplines_hint": "Generate demo data or use the API to seed events.",
+    "event_detail.last_update": "Updated",
+    "event_detail.status.scheduled": "Scheduled",
+    "event_detail.status.live": "Live",
+    "event_detail.status.completed": "Completed",
+    "event_detail.status.delayed": "Delayed",
+    "event_detail.status.finalized": "Finalized",
+    "event_detail.status.ready": "Ready",
+    "event_detail.status.finished": "Finished",
+    "event_detail.status.dns": "Did not start",
+    "event_detail.status.dq": "Disqualified",
+    "event_detail.table_position": "Pos",
+    "event_detail.table_lane": "Lane",
+    "event_detail.table_athlete": "Athlete",
+    "event_detail.table_team": "Team / Roster",
+    "event_detail.table_result": "Result",
+    "event_detail.table_points": "Points",
+    "event_detail.table_status": "Status",
+    "event_detail.meta.sessions": "{count} sessions scheduled",
+    "event_detail.meta.sessions_single": "1 session scheduled",
+    "event_detail.meta.live_sessions": "{count} session live",
+    "event_detail.meta.live_sessions_plural": "{count} sessions live",
+    "event_detail.generate_toast": "Demo timeline generated for this meet.",
+    "event_detail.demo_error": "Unable to generate the demo timeline.",
+    "event_detail.refresh_error": "Unable to load live event data.",
+    "event_detail.group_unscheduled": "Unassigned session",
+    "event_detail.scoreboard_empty": "No entries yet",
+    "home.rosters_title": "Rosters",
+    "home.rosters_subtitle": "Keep squads aligned with verified athlete eligibility.",
+    "home.rosters_empty": "No rosters available yet.",
+    "home.rosters_hint": "Federations upload rosters directly for instant publication.",
+    "home.news_title": "News",
+    "home.news_subtitle": "Bilingual coverage powered by Trackeo correspondents and partners.",
+    "home.news_empty": "No news stories have been published yet.",
+    "home.news_hint": "Follow Trackeo Insights for the latest headlines.",
+    "premium.title": "Coach & Federation tiers",
+    "premium.subtitle": "Unlock deep analytics, heat maps, and race video archives tailored to your role.",
+    "premium.compare": "Compare plans",
+    "premium.guest_title": "Guest",
+    "premium.guest_price": "Free",
+    "premium.guest_benefit_1": "Open event calendar",
+    "premium.guest_benefit_2": "Headline stats and news highlights",
+    "premium.guest_benefit_3": "Regional localization (ES/PT)",
+    "premium.premium_title": "Premium",
+    "premium.premium_price": "$9 / month",
+    "premium.premium_benefit_1": "Full athlete history & season analytics",
+    "premium.premium_benefit_2": "Video library with race markers",
+    "premium.premium_benefit_3": "Priority support in English & Spanish",
+    "premium.coach_title": "Coach",
+    "premium.coach_price": "$19 / month",
+    "premium.coach_benefit_1": "Roster compliance dashboard",
+    "premium.coach_benefit_2": "Recruiting heat maps",
+    "premium.coach_benefit_3": "Direct federation channels",
+    "premium.federation_title": "Federation",
+    "premium.federation_price": "$99 / month",
+    "premium.federation_benefit_1": "Secure data uploads",
+    "premium.federation_benefit_2": "Automated verification",
+    "premium.federation_benefit_3": "Ops support & translations",
+    "profiles.title": "Athlete & coach profiles",
+    "profiles.subtitle": "Review verified bios and connect with top performers.",
+    "profiles.refresh": "Refresh",
+    "profiles.create": "Create profile",
+    "profiles.empty": "No profiles available yet.",
+    "profiles.filter_placeholder": "Filter by name or email",
+    "profiles.filter_aria": "Filter profiles",
+    "events.title": "Event calendar",
+    "events.subtitle": "Monitor competitions across the Trackeo network.",
+    "events.refresh": "Refresh",
+    "events.create": "Create event",
+    "events.upcoming": "Only show upcoming meets",
+    "events.empty": "No events published yet.",
+    "rosters.title": "Roster library",
+    "rosters.subtitle": "Track lineup updates and eligibility.",
+    "rosters.refresh": "Refresh",
+    "rosters.filter_placeholder": "Filter by club or country",
+    "rosters.filter_aria": "Filter rosters",
+    "rosters.empty": "No rosters available yet.",
+    "rosters.hint": "Federations can submit files via the secure area.",
+    "rosters.list_title": "Latest submissions",
+    "rosters.list_subtitle": "Verified rosters arrive via secure federation uploads.",
+    "rosters.list_empty": "No submissions yet.",
+    "athlete_detail.history_empty": "No race history uploaded yet.",
+    "athlete_detail.rosters_empty": "No rosters attached yet.",
+    "roster_detail.athletes_empty": "No athletes listed yet.",
+    "login.title": "Sign in to Trackeo",
+    "login.subtitle": "Access premium tools with your verified email.",
+    "login.submit": "Log in",
+    "login.switch": "Need an account?",
+    "login.switch_link": "Create one now.",
+    "signup.title": "Create your Trackeo account",
+    "signup.subtitle": "Manage rosters, follow events, and unlock insights.",
+    "signup.tier": "Subscription plan",
+    "signup.tier_free": "Free",
+    "signup.tier_premium": "Premium",
+    "signup.tier_coach": "Coach",
+    "signup.tier_federation": "Federation",
+    "signup.role_fan": "Fan",
+    "signup.role_athlete": "Athlete",
+    "signup.role_coach": "Coach",
+    "signup.role_federation": "Federation",
+    "signup.role_scout": "Scout",
+    "signup.submit": "Create account",
+    "signup.switch": "Already have an account?",
+    "signup.switch_link": "Sign in.",
+    "federations.title": "Federation secure uploads",
+    "federations.subtitle": "Submit encrypted payloads with audit trails and automatic validation.",
+    "federations.refresh": "Refresh submissions",
+    "federations.form_title": "Upload official files",
+    "federations.form_subtitle":
+      "Provide a signed link. Trackeo validates checksums and processes asynchronously.",
+    "federations.token": "Access token",
+    "federations.token_placeholder": "Bearer token",
+    "federations.name": "Federation name",
+    "federations.name_placeholder": "South American Confederation",
+    "federations.email_placeholder": "contact@federation.org",
+    "federations.payload": "Secure package URL",
+    "federations.payload_placeholder": "https://storage.example.com/results.json",
+    "federations.notes": "Notes (optional)",
+    "federations.notes_placeholder": "Describe the meet or include validation details",
+    "federations.submit": "Submit for processing",
+    "federations.submissions_title": "Submission history",
+    "federations.submissions_subtitle": "Track processing, checksum, and verification.",
+    "federations.submissions_empty": "No submissions yet. Upload the first official file.",
+    "about.title": "About Trackeo",
+    "about.subtitle": "Built in Atlanta and powered by Latin American federations.",
+    "about.mission_title": "Mission",
+    "about.mission_subtitle":
+      "We amplify athletes with trusted data, multilingual coverage, and equitable access.",
+    "about.mission_body_1":
+      "Trackeo centralizes competition data from 19 federations and delivers insights in English, Spanish, and Portuguese.",
+    "about.mission_body_2":
+      "We support coaches with analytics, engage fans with stories, and simplify compliance for federation teams.",
+    "about.team_title": "Team & partners",
+    "about.team_subtitle": "We collaborate with statisticians, journalists, and federations across the region.",
+    "about.team_item_1_title": "Atlanta HQ",
+    "about.team_item_1_body": "Product, engineering, and partnerships operate from Atlanta coordinating multilingual coverage.",
+    "about.team_item_2_title": "Regional bureaus",
+    "about.team_item_2_body": "Correspondents in Bogotá, São Paulo, Santiago, and Mexico City verify results and produce news.",
+    "about.team_item_3_title": "Federation council",
+    "about.team_item_3_body": "An advisory board ensures Trackeo meets security and accessibility standards.",
+  },
+  es: {
+    "badge.beta": "Beta",
+    "nav.home": "Inicio",
+    "nav.profiles": "Perfiles",
+    "nav.events": "Eventos",
+    "nav.rosters": "Planteles",
+    "nav.federations": "Federaciones",
+    "nav.about": "Acerca de",
+    "locale.label": "Elige idioma",
+    "action.login": "Iniciar sesión",
+    "action.signup": "Crear cuenta",
+    "footer.powered": "Trackeo · Impulsado por FastAPI · Base API:",
+    "footer.location": "Con sede en Atlanta, GA · Escalando por Sudamérica con socios locales.",
+    "hero.eyebrow": "Inteligencia del atletismo latinoamericano",
+    "hero.title": "Trackeo conecta federaciones, entrenadores y fans en las Américas.",
+    "hero.description":
+      "Explora marcas verificadas, noticias multilingües y análisis premium que resaltan el auge del atletismo de São Paulo a Bogotá. Construido en Atlanta para el continente.",
+    "hero.explore": "Explorar como invitado",
+    "hero.subscribe": "Ver planes para entrenadores",
+    "preview.title": "Destacado regional",
+    "preview.subtitle": "19 federaciones transmitiendo parciales y heat sheets en vivo.",
+    "preview.metric_meets": "Eventos verificados",
+    "preview.metric_athletes": "Atletas seguidos",
+    "preview.metric_rosters": "Actualizaciones de planteles",
+    "search.title": "Busca en el universo Trackeo",
+    "search.subtitle": "Encuentra atletas, eventos, planteles y noticias bilingües en un solo lugar.",
+    "search.placeholder": "Busca un atleta, club o evento",
+    "search.aria": "Buscar en Trackeo",
+    "search.filters": "Categorías de búsqueda",
+    "search.filter_all": "Todos",
+    "search.filter_athletes": "Atletas",
+    "search.filter_events": "Eventos",
+    "search.filter_rosters": "Planteles",
+    "search.filter_news": "Noticias",
+    "search.empty": "Empieza a escribir para explorar el universo de datos Trackeo.",
+    "search.no_results": "Aún no hay coincidencias. Ajusta filtros o revisa la ortografía.",
+    "home.athletes_title": "Atletas",
+    "home.athletes_subtitle": "Explora el plantel o destaca una promesa de tu federación.",
+    "home.athletes_seed": "Recargar atletas de ejemplo",
+    "home.athletes_form_title": "Crear perfil de atleta",
+    "home.athletes_submit": "Registrar atleta",
+    "home.athletes_empty": "Aún no se registran atletas.",
+    "home.athletes_hint": "Usa el formulario o carga atletas de ejemplo para poblar el plantel.",
+    "home.events_title": "Eventos",
+    "home.events_subtitle": "Sigue las competencias de Atlanta a Buenos Aires.",
+    "home.events_seed": "Recargar eventos de ejemplo",
+    "home.events_form_title": "Programar nuevo evento",
+    "home.events_submit": "Crear evento",
+    "home.events_empty": "Aún no se programan eventos.",
+    "home.events_hint": "Usa el formulario o recarga el calendario curado de eventos.",
+    "event_detail.eyebrow": "Operaciones en vivo",
+    "event_detail.refresh": "Actualizar",
+    "event_detail.generate_demo": "Generar cronograma demo",
+    "event_detail.overview_title": "Resumen del evento",
+    "event_detail.overview_subtitle": "Estado, federación y última actualización.",
+    "event_detail.summary_location": "Ubicación",
+    "event_detail.summary_dates": "Fechas",
+    "event_detail.summary_federation": "Federación",
+    "event_detail.summary_status": "Estado en vivo",
+    "event_detail.schedule_title": "Sesiones y cronograma",
+    "event_detail.schedule_subtitle": "Monitorea bloques de competencia y sedes.",
+    "event_detail.schedule_empty": "Aún no hay sesiones programadas.",
+    "event_detail.schedule_hint": "Crea una sesión o genera datos demo para poblar la línea de tiempo.",
+    "event_detail.disciplines_title": "Pruebas y marcadores",
+    "event_detail.disciplines_subtitle": "Las heat sheets se actualizan al instante en pista.",
+    "event_detail.disciplines_empty": "No hay pruebas asociadas a este evento.",
+    "event_detail.disciplines_hint": "Genera datos demo o usa la API para cargar eventos.",
+    "event_detail.last_update": "Actualizado",
+    "event_detail.status.scheduled": "Programado",
+    "event_detail.status.live": "En vivo",
+    "event_detail.status.completed": "Completado",
+    "event_detail.status.delayed": "Demorado",
+    "event_detail.status.finalized": "Finalizado",
+    "event_detail.status.ready": "Listo",
+    "event_detail.status.finished": "Terminado",
+    "event_detail.status.dns": "No largó",
+    "event_detail.status.dq": "Descalificado",
+    "event_detail.table_position": "Pos",
+    "event_detail.table_lane": "Carril",
+    "event_detail.table_athlete": "Atleta",
+    "event_detail.table_team": "Equipo / Plantel",
+    "event_detail.table_result": "Resultado",
+    "event_detail.table_points": "Puntos",
+    "event_detail.table_status": "Estado",
+    "event_detail.meta.sessions": "{count} sesiones programadas",
+    "event_detail.meta.sessions_single": "1 sesión programada",
+    "event_detail.meta.live_sessions": "{count} sesión en vivo",
+    "event_detail.meta.live_sessions_plural": "{count} sesiones en vivo",
+    "event_detail.generate_toast": "Cronograma demo generado para este evento.",
+    "event_detail.demo_error": "No se pudo generar el cronograma demo.",
+    "event_detail.refresh_error": "No se pudo cargar la información en vivo.",
+    "event_detail.group_unscheduled": "Sesión sin asignar",
+    "event_detail.scoreboard_empty": "Sin entradas",
+    "home.rosters_title": "Planteles",
+    "home.rosters_subtitle": "Mantén las escuadras en regla con elegibilidad verificada.",
+    "home.rosters_empty": "Aún no hay planteles.",
+    "home.rosters_hint": "Las federaciones cargan planteles para publicación instantánea.",
+    "home.news_title": "Noticias",
+    "home.news_subtitle": "Cobertura bilingüe con corresponsales y socios Trackeo.",
+    "home.news_empty": "Aún no hay noticias publicadas.",
+    "home.news_hint": "Sigue Trackeo Insights para las últimas novedades.",
+    "premium.title": "Planes para coaches y federaciones",
+    "premium.subtitle": "Desbloquea analíticas, mapas y archivos de video según tu rol.",
+    "premium.compare": "Comparar planes",
+    "premium.guest_title": "Invitado",
+    "premium.guest_price": "Gratis",
+    "premium.guest_benefit_1": "Calendario abierto",
+    "premium.guest_benefit_2": "Estadísticas destacadas y noticias",
+    "premium.guest_benefit_3": "Localización regional (ES/PT)",
+    "premium.premium_title": "Premium",
+    "premium.premium_price": "US$9 / mes",
+    "premium.premium_benefit_1": "Historial completo y análisis",
+    "premium.premium_benefit_2": "Videoteca con marcadores",
+    "premium.premium_benefit_3": "Soporte prioritario en inglés y español",
+    "premium.coach_title": "Coach",
+    "premium.coach_price": "US$19 / mes",
+    "premium.coach_benefit_1": "Tablero de elegibilidad",
+    "premium.coach_benefit_2": "Mapas de scouting",
+    "premium.coach_benefit_3": "Canales directos con federaciones",
+    "premium.federation_title": "Federación",
+    "premium.federation_price": "US$99 / mes",
+    "premium.federation_benefit_1": "Cargas seguras",
+    "premium.federation_benefit_2": "Verificación automática",
+    "premium.federation_benefit_3": "Soporte operativo y traducciones",
+    "profiles.title": "Perfiles de atletas y coaches",
+    "profiles.subtitle": "Revisa bios verificadas y conecta con protagonistas.",
+    "profiles.refresh": "Actualizar",
+    "profiles.create": "Crear perfil",
+    "profiles.empty": "Aún no hay perfiles.",
+    "profiles.filter_placeholder": "Filtrar por nombre o correo",
+    "profiles.filter_aria": "Filtrar perfiles",
+    "events.title": "Calendario de eventos",
+    "events.subtitle": "Monitorea competencias en la red Trackeo.",
+    "events.refresh": "Actualizar",
+    "events.create": "Crear evento",
+    "events.upcoming": "Solo mostrar próximos eventos",
+    "events.empty": "Aún no hay eventos publicados.",
+    "rosters.title": "Biblioteca de planteles",
+    "rosters.subtitle": "Sigue actualizaciones de alineaciones y elegibilidad.",
+    "rosters.refresh": "Actualizar",
+    "rosters.filter_placeholder": "Filtrar por club o país",
+    "rosters.filter_aria": "Filtrar planteles",
+    "rosters.empty": "Aún no hay planteles.",
+    "rosters.hint": "Las federaciones pueden enviar archivos en el área segura.",
+    "rosters.list_title": "Últimos envíos",
+    "rosters.list_subtitle": "Planteles verificados llegan por cargas seguras de federaciones.",
+    "rosters.list_empty": "Sin envíos aún.",
+    "athlete_detail.history_empty": "Aún no hay historial de competencias.",
+    "athlete_detail.rosters_empty": "Aún no hay planteles asociados.",
+    "roster_detail.athletes_empty": "Aún no hay atletas listados.",
+    "login.title": "Inicia sesión en Trackeo",
+    "login.subtitle": "Accede a las herramientas premium con tu correo verificado.",
+    "login.submit": "Iniciar sesión",
+    "login.switch": "¿Necesitas una cuenta?",
+    "login.switch_link": "Créala ahora.",
+    "signup.title": "Crea tu cuenta Trackeo",
+    "signup.subtitle": "Gestiona planteles, sigue eventos y obtén análisis.",
+    "signup.tier": "Plan de suscripción",
+    "signup.tier_free": "Gratis",
+    "signup.tier_premium": "Premium",
+    "signup.tier_coach": "Coach",
+    "signup.tier_federation": "Federación",
+    "signup.role_fan": "Fan",
+    "signup.role_athlete": "Atleta",
+    "signup.role_coach": "Coach",
+    "signup.role_federation": "Federación",
+    "signup.role_scout": "Scout",
+    "signup.submit": "Crear cuenta",
+    "signup.switch": "¿Ya tienes cuenta?",
+    "signup.switch_link": "Inicia sesión.",
+    "federations.title": "Cargas seguras de federaciones",
+    "federations.subtitle":
+      "Envía paquetes cifrados con trazabilidad y verificación automática.",
+    "federations.refresh": "Actualizar envíos",
+    "federations.form_title": "Enviar archivos oficiales",
+    "federations.form_subtitle":
+      "Proporciona un enlace firmado. Trackeo valida checksum y procesa de forma asíncrona.",
+    "federations.token": "Token de acceso",
+    "federations.token_placeholder": "Token Bearer",
+    "federations.name": "Nombre de la federación",
+    "federations.name_placeholder": "Confederación Sudamericana",
+    "federations.email_placeholder": "contacto@federacion.org",
+    "federations.payload": "URL segura del paquete",
+    "federations.payload_placeholder": "https://storage.ejemplo.com/resultados.json",
+    "federations.notes": "Notas (opcional)",
+    "federations.notes_placeholder": "Describe el evento o incluye detalles de validación",
+    "federations.submit": "Enviar a procesamiento",
+    "federations.submissions_title": "Historial de envíos",
+    "federations.submissions_subtitle": "Monitorea procesamiento, checksum y verificación.",
+    "federations.submissions_empty": "Sin envíos aún. Carga el primer archivo oficial.",
+    "about.title": "Sobre Trackeo",
+    "about.subtitle": "Construido en Atlanta e impulsado por federaciones latinoamericanas.",
+    "about.mission_title": "Misión",
+    "about.mission_subtitle":
+      "Amplificamos atletas con datos confiables, cobertura multilingüe y acceso equitativo.",
+    "about.mission_body_1":
+      "Trackeo centraliza datos de 19 federaciones y entrega insights en inglés, español y portugués.",
+    "about.mission_body_2":
+      "Apoyamos coaches con analíticas, involucramos fans con historias y simplificamos la conformidad para equipos federativos.",
+    "about.team_title": "Equipo y socios",
+    "about.team_subtitle": "Colaboramos con estadísticos, periodistas y federaciones en toda la región.",
+    "about.team_item_1_title": "Sede Atlanta",
+    "about.team_item_1_body": "Producto, ingeniería y partnerships operan en Atlanta coordinando cobertura multilingüe.",
+    "about.team_item_2_title": "Bureaus regionales",
+    "about.team_item_2_body": "Corresponsales en Bogotá, São Paulo, Santiago y Ciudad de México verifican resultados y producen noticias.",
+    "about.team_item_3_title": "Consejo federativo",
+    "about.team_item_3_body": "Un consejo consultivo asegura que Trackeo cumpla estándares de seguridad y accesibilidad.",
+  },
+  pt: {
+    "badge.beta": "Beta",
+    "nav.home": "Início",
+    "nav.profiles": "Perfis",
+    "nav.events": "Eventos",
+    "nav.rosters": "Elencos",
+    "nav.federations": "Federações",
+    "nav.about": "Sobre",
+    "locale.label": "Escolha o idioma",
+    "action.login": "Entrar",
+    "action.signup": "Criar conta",
+    "footer.powered": "Trackeo · Impulsionado por FastAPI · Base da API:",
+    "footer.location": "Sede em Atlanta, GA · Expandindo pela América do Sul com parceiros locais.",
+    "hero.eyebrow": "Inteligência do atletismo latino-americano",
+    "hero.title": "Trackeo conecta federações, técnicos e fãs pelas Américas.",
+    "hero.description":
+      "Explore performances verificadas, notícias multilíngues e insights premium que destacam o atletismo de São Paulo a Bogotá. Construído em Atlanta para o continente.",
+    "hero.explore": "Explorar como convidado",
+    "hero.subscribe": "Ver planos para técnicos",
+    "preview.title": "Destaque regional",
+    "preview.subtitle": "19 federações transmitindo parciais e heat sheets ao vivo.",
+    "preview.metric_meets": "Eventos verificados",
+    "preview.metric_athletes": "Atletas monitorados",
+    "preview.metric_rosters": "Atualizações de elencos",
+    "search.title": "Busque no universo Trackeo",
+    "search.subtitle": "Encontre atletas, eventos, elencos e notícias bilíngues em um só lugar.",
+    "search.placeholder": "Busque por atleta, clube ou competição",
+    "search.aria": "Buscar no Trackeo",
+    "search.filters": "Categorias de busca",
+    "search.filter_all": "Todos",
+    "search.filter_athletes": "Atletas",
+    "search.filter_events": "Eventos",
+    "search.filter_rosters": "Elencos",
+    "search.filter_news": "Notícias",
+    "search.empty": "Comece a digitar para explorar o universo Trackeo.",
+    "search.no_results": "Nenhum resultado ainda. Ajuste filtros ou ortografia.",
+    "home.athletes_title": "Atletas",
+    "home.athletes_subtitle": "Explore o elenco ou destaque um talento da sua federação.",
+    "home.athletes_seed": "Recarregar atletas de exemplo",
+    "home.athletes_form_title": "Criar perfil de atleta",
+    "home.athletes_submit": "Registrar atleta",
+    "home.athletes_empty": "Nenhum atleta cadastrado ainda.",
+    "home.athletes_hint": "Use o formulário ou carregue atletas de exemplo para preencher o elenco.",
+    "home.events_title": "Eventos",
+    "home.events_subtitle": "Acompanhe competições de Atlanta a Buenos Aires.",
+    "home.events_seed": "Recarregar eventos de exemplo",
+    "home.events_form_title": "Agendar novo evento",
+    "home.events_submit": "Criar evento",
+    "home.events_empty": "Nenhum evento agendado ainda.",
+    "home.events_hint": "Use o formulário ou recarregue o calendário curado.",
+    "event_detail.eyebrow": "Operações ao vivo",
+    "event_detail.refresh": "Atualizar",
+    "event_detail.generate_demo": "Gerar cronograma demo",
+    "event_detail.overview_title": "Visão geral do evento",
+    "event_detail.overview_subtitle": "Acompanhe status, federação e última atualização.",
+    "event_detail.summary_location": "Local",
+    "event_detail.summary_dates": "Datas",
+    "event_detail.summary_federation": "Federação",
+    "event_detail.summary_status": "Status ao vivo",
+    "event_detail.schedule_title": "Sessões e cronograma",
+    "event_detail.schedule_subtitle": "Monitore blocos de competição e locais.",
+    "event_detail.schedule_empty": "Nenhuma sessão agendada ainda.",
+    "event_detail.schedule_hint": "Crie uma sessão ou gere dados demo para preencher a linha do tempo.",
+    "event_detail.disciplines_title": "Provas e placares",
+    "event_detail.disciplines_subtitle": "As heat sheets atualizam instantaneamente à beira da pista.",
+    "event_detail.disciplines_empty": "Nenhuma prova vinculada a este evento.",
+    "event_detail.disciplines_hint": "Gere dados demo ou use a API para cadastrar provas.",
+    "event_detail.last_update": "Atualizado",
+    "event_detail.status.scheduled": "Programado",
+    "event_detail.status.live": "Ao vivo",
+    "event_detail.status.completed": "Concluído",
+    "event_detail.status.delayed": "Atrasado",
+    "event_detail.status.finalized": "Finalizado",
+    "event_detail.status.ready": "Pronto",
+    "event_detail.status.finished": "Encerrado",
+    "event_detail.status.dns": "Não largou",
+    "event_detail.status.dq": "Desclassificado",
+    "event_detail.table_position": "Pos",
+    "event_detail.table_lane": "Raia",
+    "event_detail.table_athlete": "Atleta",
+    "event_detail.table_team": "Equipe / Elenco",
+    "event_detail.table_result": "Resultado",
+    "event_detail.table_points": "Pontos",
+    "event_detail.table_status": "Status",
+    "event_detail.meta.sessions": "{count} sessões agendadas",
+    "event_detail.meta.sessions_single": "1 sessão agendada",
+    "event_detail.meta.live_sessions": "{count} sessão ao vivo",
+    "event_detail.meta.live_sessions_plural": "{count} sessões ao vivo",
+    "event_detail.generate_toast": "Cronograma demo gerado para este evento.",
+    "event_detail.demo_error": "Não foi possível gerar o cronograma demo.",
+    "event_detail.refresh_error": "Não foi possível carregar dados ao vivo.",
+    "event_detail.group_unscheduled": "Sessão sem atribuição",
+    "event_detail.scoreboard_empty": "Sem entradas",
+    "home.rosters_title": "Elencos",
+    "home.rosters_subtitle": "Mantenha as equipes alinhadas com elegibilidade verificada.",
+    "home.rosters_empty": "Nenhum elenco disponível ainda.",
+    "home.rosters_hint": "Federações enviam elencos diretamente para publicação instantânea.",
+    "home.news_title": "Notícias",
+    "home.news_subtitle": "Cobertura bilíngue com correspondentes e parceiros Trackeo.",
+    "home.news_empty": "Nenhuma notícia publicada ainda.",
+    "home.news_hint": "Siga Trackeo Insights para as últimas manchetes.",
+    "premium.title": "Planos para técnicos e federações",
+    "premium.subtitle": "Desbloqueie análises, mapas de calor e arquivos de vídeo conforme seu papel.",
+    "premium.compare": "Comparar planos",
+    "premium.guest_title": "Convidado",
+    "premium.guest_price": "Grátis",
+    "premium.guest_benefit_1": "Calendário aberto",
+    "premium.guest_benefit_2": "Estatísticas e destaques",
+    "premium.guest_benefit_3": "Localização regional (ES/PT)",
+    "premium.premium_title": "Premium",
+    "premium.premium_price": "US$9 / mês",
+    "premium.premium_benefit_1": "Histórico completo e análises",
+    "premium.premium_benefit_2": "Videoteca com marcadores",
+    "premium.premium_benefit_3": "Suporte prioritário em inglês e espanhol",
+    "premium.coach_title": "Coach",
+    "premium.coach_price": "US$19 / mês",
+    "premium.coach_benefit_1": "Painel de conformidade",
+    "premium.coach_benefit_2": "Mapas de recrutamento",
+    "premium.coach_benefit_3": "Canais diretos com federações",
+    "premium.federation_title": "Federação",
+    "premium.federation_price": "US$99 / mês",
+    "premium.federation_benefit_1": "Uploads seguros",
+    "premium.federation_benefit_2": "Verificação automática",
+    "premium.federation_benefit_3": "Suporte operacional e traduções",
+    "profiles.title": "Perfis de atletas e técnicos",
+    "profiles.subtitle": "Revise bios verificadas e conecte-se com destaques.",
+    "profiles.refresh": "Atualizar",
+    "profiles.create": "Criar perfil",
+    "profiles.empty": "Nenhum perfil disponível ainda.",
+    "profiles.filter_placeholder": "Filtrar por nome ou e-mail",
+    "profiles.filter_aria": "Filtrar perfis",
+    "events.title": "Calendário de eventos",
+    "events.subtitle": "Monitore competições na rede Trackeo.",
+    "events.refresh": "Atualizar",
+    "events.create": "Criar evento",
+    "events.upcoming": "Mostrar apenas próximos eventos",
+    "events.empty": "Nenhum evento publicado ainda.",
+    "rosters.title": "Biblioteca de elencos",
+    "rosters.subtitle": "Acompanhe atualizações e elegibilidade.",
+    "rosters.refresh": "Atualizar",
+    "rosters.filter_placeholder": "Filtrar por clube ou país",
+    "rosters.filter_aria": "Filtrar elencos",
+    "rosters.empty": "Nenhum elenco disponível ainda.",
+    "rosters.hint": "Federações podem enviar arquivos na área segura.",
+    "rosters.list_title": "Últimos envios",
+    "rosters.list_subtitle": "Elencos verificados chegam por uploads seguros das federações.",
+    "rosters.list_empty": "Nenhum envio ainda.",
+    "athlete_detail.history_empty": "Nenhum histórico de provas cadastrado.",
+    "athlete_detail.rosters_empty": "Nenhum elenco associado ainda.",
+    "roster_detail.athletes_empty": "Nenhum atleta listado ainda.",
+    "login.title": "Faça login no Trackeo",
+    "login.subtitle": "Use seu e-mail verificado para acessar as ferramentas premium.",
+    "login.submit": "Entrar",
+    "login.switch": "Precisa de uma conta?",
+    "login.switch_link": "Crie agora.",
+    "signup.title": "Crie sua conta Trackeo",
+    "signup.subtitle": "Junte-se para gerenciar elencos, acompanhar eventos e obter análises.",
+    "signup.tier": "Plano de assinatura",
+    "signup.tier_free": "Gratuito",
+    "signup.tier_premium": "Premium",
+    "signup.tier_coach": "Coach",
+    "signup.tier_federation": "Federação",
+    "signup.role_fan": "Fã",
+    "signup.role_athlete": "Atleta",
+    "signup.role_coach": "Coach",
+    "signup.role_federation": "Federação",
+    "signup.role_scout": "Olheiro",
+    "signup.submit": "Criar conta",
+    "signup.switch": "Já possui cadastro?",
+    "signup.switch_link": "Faça login.",
+    "federations.title": "Uploads seguros das federações",
+    "federations.subtitle":
+      "Envie pacotes criptografados com trilhas de auditoria e verificação automática.",
+    "federations.refresh": "Atualizar envios",
+    "federations.form_title": "Enviar arquivos oficiais",
+    "federations.form_subtitle":
+      "Forneça um link assinado. A Trackeo valida o checksum e processa de forma assíncrona.",
+    "federations.token": "Token de acesso",
+    "federations.token_placeholder": "Token Bearer",
+    "federations.name": "Nome da federação",
+    "federations.name_placeholder": "Confederação Sul-Americana",
+    "federations.email_placeholder": "contato@federacao.org",
+    "federations.payload": "URL segura do pacote",
+    "federations.payload_placeholder": "https://storage.exemplo.com/resultados.json",
+    "federations.notes": "Notas (opcional)",
+    "federations.notes_placeholder": "Descreva o evento ou inclua detalhes de validação",
+    "federations.submit": "Enviar para processamento",
+    "federations.submissions_title": "Histórico de envios",
+    "federations.submissions_subtitle": "Acompanhe processamento, checksum e verificação.",
+    "federations.submissions_empty": "Nenhum envio ainda. Faça upload do primeiro arquivo oficial.",
+    "about.title": "Sobre a Trackeo",
+    "about.subtitle": "Construída em Atlanta e impulsionada por federações da América Latina.",
+    "about.mission_title": "Missão",
+    "about.mission_subtitle":
+      "Amplificamos atletas com dados confiáveis, cobertura multilíngue e acesso equitativo.",
+    "about.mission_body_1":
+      "A Trackeo centraliza dados de competição de 19 federações e entrega insights em inglês, espanhol e português.",
+    "about.mission_body_2":
+      "Apoiamos técnicos com análises, engajamos fãs com histórias e simplificamos a conformidade para equipes federativas.",
+    "about.team_title": "Equipe e parceiros",
+    "about.team_subtitle": "Colaboramos com estatísticos, jornalistas e federações em toda a região.",
+    "about.team_item_1_title": "Sede Atlanta",
+    "about.team_item_1_body": "Produto, engenharia e parcerias operam em Atlanta coordenando cobertura multilíngue.",
+    "about.team_item_2_title": "Bureaus regionais",
+    "about.team_item_2_body": "Correspondentes em Bogotá, São Paulo, Santiago e Cidade do México verificam resultados e produzem notícias.",
+    "about.team_item_3_title": "Conselho federativo",
+    "about.team_item_3_body": "Um conselho consultivo garante que a Trackeo atenda padrões de segurança e acessibilidade.",
+  },
+};
+
+export function applyLocale(locale, { switcher } = {}) {
+  const resolved = translations[locale] ? locale : "en";
+  const dictionary = translations[resolved];
+  document.documentElement.lang = resolved;
+  document.querySelectorAll("[data-l10n-key]").forEach((element) => {
+    const key = element.dataset.l10nKey;
+    if (key in dictionary) {
+      element.textContent = dictionary[key];
+    }
+  });
+  document.querySelectorAll("[data-l10n-placeholder]").forEach((element) => {
+    const key = element.dataset.l10nPlaceholder;
+    if (key in dictionary) {
+      element.setAttribute("placeholder", dictionary[key]);
+    }
+  });
+  document.querySelectorAll("[data-l10n-aria-label]").forEach((element) => {
+    const key = element.dataset.l10nAriaLabel;
+    if (key in dictionary) {
+      element.setAttribute("aria-label", dictionary[key]);
+    }
+  });
+  if (switcher) {
+    switcher.value = resolved;
+  }
+  try {
+    localStorage.setItem("trackeo.locale", resolved);
+  } catch (error) {
+    // ignore storage issues
+  }
+  return resolved;
+}
+
+export function readStoredLocale() {
+  try {
+    return localStorage.getItem("trackeo.locale");
+  } catch (error) {
+    return null;
+  }
+}
